% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/authority2csv.R
\name{authority2csv}
\alias{authority2csv}
\title{Convert Author-ity data to .csv}
\usage{
authority2csv(files, file, no.lines = 10000, wh.col = 1:25, cl)
}
\arguments{
\item{files}{Character vector; names of Author-ity files to include}

\item{file}{Character singleton; name of destination `.csv` file}

\item{no.lines}{Numeric; number of lines to read from `files` at a time}

\item{wh.col}{Numeric; which columns of Author-ity to include in `file`}

\item{cl}{Cluster; optional}
}
\description{
This function converts specified elements of Author-ity data (for main
results, not those using only 10 most recent papers) to a `.csv` file.
}

