% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/count.instances.R
\name{count.instances}
\alias{count.instances}
\title{Count instances of a string}
\usage{
count.instances(x, pattern)
}
\arguments{
\item{x}{Vector or list of strings (to be fed to `strsplit`)}

\item{pattern}{Character string}
}
\description{
This function counts the numbers of instances of a string in each of a vector
or list of strings.
}

