% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{file.rans}
\alias{file.rans}
\title{Begin and end dates for the PubMed files}
\format{A data.frame with columns `begin` and `end`.}
\source{
Xintao Wei
}
\description{
A data.frame of the first and last year for which some publication from that
year appears in each of the 684 PubMed data files.
}

