% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hel2bigraph.R
\name{hel2bigraph}
\alias{hel2bigraph}
\title{Create a bigraph from a hyperedge list}
\usage{
hel2bigraph(hel, type = FALSE)
}
\arguments{
\item{hel}{data.table with first column one type of node and second column a
list of nodes of the other type tied to the corresponding node in the first.}

\item{type}{Logical; node type for the first column.}
}
\description{
This function constructs a bipartite graph from a hyperedge list given as a
data.table.
}

