% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{jids}
\alias{jids}
\title{PubMed and NCBI Molecular Biology Database Journals}
\format{A data.table.}
\source{
National Library of Medicine
\url{http://www.ncbi.nlm.nih.gov/books/NBK3827/table/pubmedhelp.pubmedhelptable45/}
}
\description{
A data.table of PubMed and NCBI Journal identifiers with full and abbreviated
journal titles and print and online ISSN numbers.
}
\details{
A data.table of 2014 MeSH terms and their hierarchical tree numbers. This
object is adapted from the source file `mtrees2014.bin` using the script
`mtrees.R`.
}

