% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mesh.family.R
\name{mesh.family}
\alias{mesh.family}
\title{Collect MeSH terms at or below a given set of parents}
\usage{
mesh.family(vec)
}
\arguments{
\item{vec}{Character vector; parent MeSH terms whose children to include}
}
\description{
This function takes a vector of parent MeSH terms and returns a vector of
parent and children MeSH terms.
}

