% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{mtrees}
\alias{mtrees}
\title{2014 Medical Subject Headings tree}
\format{A data.table with first column the MeSH terms and second column the
corresponding hierarchical tree numbers.}
\source{
National Library of Medicine
\url{http://www.nlm.nih.gov/mesh/filelist.html}
}
\description{
A data.table of 2014 MeSH terms and their hierarchical tree numbers. This
object is adapted from the source file `mtrees2014.bin` using the script
`mtrees.R`.
}

