% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subsplit.R
\name{subsplit}
\alias{subsplit}
\title{Split lines and keep only certain entries}
\usage{
subsplit(x, pattern, inds = 1, cl, parallel = !missing(cl))
}
\arguments{
\item{x}{Vector or list of strings (to be fed to `strsplit`)}

\item{pattern}{Character string}

\item{inds}{Numerical vector; entries of each `strsplit` to keep}
}
\description{
This function splits a list or vector of lines by a pattern but includes only
certain entries in the output (all of which must be present in each split
string).
}

