% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Template function file}
\usage{
template(x)
}
\arguments{
\item{x}{Arbitrary; nothing is done with it.}
}
\description{
This function does nothing; the file exists to be copied and edited into new
function files.
}

