% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/year.files.R
\name{year.files}
\alias{year.files}
\title{File numbers for a given year}
\usage{
year.files(year)
}
\arguments{
\item{year}{Numeric; a year.}
}
\description{
This function identifies the numbers of the PubMed files containing entries
dated to a given year.
}

