% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.pubmed.R
\name{read.pubmed}
\alias{read.pubmed}
\title{Read PubMed data to a data frame}
\usage{
read.pubmed(file, nrows = -1, skip = 0)
}
\arguments{
\item{file}{Character string; the name of the data file}

\item{nrows}{Numeric; the number of rows to read}

\item{skip}{Numeric; the number of lines to skip}
}
\description{
This function reads PubMed data to a data frame.
}

